+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 - Added github link to modDesc.xml
 - Refactored TempInfo.lua
-- Decreased TempInfo min/max temperatur font size
+- Decreased TempInfo min/max temperatur font size - for #2
+- Added Settings area to disable temp info - for #1
 
 ## [1.0.0.0] - 2023-04-17
 
-### Added or Changed
+### Added
 - Initial release (https://www.farming-simulator.com/mod.php?mod_id=267536)